requirements.txt
fastapi
httpx
uvicorn
pyyaml
pydantic
